% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Calendar.R
\name{\%+\%}
\alias{\%+\%}
\alias{nextbiz}
\alias{prevbiz}
\title{Adjusts the given dates to the next/previous business day}
\usage{
lhs \%+\% rhs

nextbiz(lhs, method = "previous", cal = getOption("calendar"))

prevbiz(lhs, method = "next", cal = getOption("calendar"))
}
\arguments{
\item{lhs}{dates to be adjusted}

\item{rhs}{offset days}

\item{cal}{an instance of \code{Calendar}}
}
\value{
\code{Date} objects adjusted accordingly.
}
\description{
If the given dates are business days it returns the given dates, but once it
is not, it returns the next/previous business days.

next biz date

previous biz date
}
\section{Date types accepted}{


The argument \code{dates} accepts \code{Date} objects and any
object that returns a valid \code{Date} object when passed through
\code{as.Date}, which include all \code{POSIX*} classes and \code{character}
objects with ISO formatted dates.
}

